Refactor figo-chat.php into a class-based handler.
- Extracted logic from `figo-chat.php` into `lib/FigoChatHandler.php`.
- Created `FigoChatHandler` class with `handle()` method and protected helper methods.
- Updated `figo-chat.php` to instantiate and use `FigoChatHandler`.
- Added unit tests in `tests/Unit/FigoChatHandlerTest.php` to verify functionality.
- Preserved existing behavior including CORS, rate limiting, and fallback logic.
